Pull functions and global variables out of interpreter
