+++ commands/Execute.md
@@ -2,11 +2,12 @@
 
 - Arity: 1
 - In: a
-- Out: exec(VY_compile(a))
+- Out: exec(vy_compile(a))
 
 Executes `a` as if it was a snippet of Vyxal code.
 
 # Usage
+
 ```
 `1 1+`ₑ  ║⟨2⟩
 `1 1+`ₑ1+║⟨3⟩
