+++ tests/test_utils.py
@@ -6,7 +6,7 @@ from multiprocessing import Manager
 THIS_FOLDER = os.path.dirname(os.path.abspath(__file__)) + "/.."
 sys.path.insert(1, THIS_FOLDER)
 
-from vyxal import interpreter
+from vyxal import interpreter as interp
 
 header = "stack = []\nregister = 0\nprinted = False\n"
 manager = Manager()
@@ -15,23 +15,23 @@ manager = Manager()
 def run_code(code, flags="", input_list=[], output_variable=manager.dict()):
     reset_globals()
     # context_level = 0
-    interpreter.execute(code, flags, "\n".join(
+    interp.execute(code, flags, "\n".join(
         map(str, input_list)), output_variable)
-    return interpreter.stack
+    return interp.stack
 
 
 def reset_globals():
-    interpreter.keg_mode = False
-    interpreter.raw_strings = False
-    interpreter.online_version = False
-    interpreter.input_level = 0
-    interpreter.number_iterable = list
-    interpreter.MAP_START = 1
-    interpreter.MAP_OFFSET = 1
-    interpreter._join = False
-    interpreter._vertical_join = False
-    interpreter.use_encoding = False
-    interpreter.stack = []
+    interp.keg_mode = False
+    interp.raw_strings = False
+    interp.online_version = False
+    interp.input_level = 0
+    interp.number_iterable = list
+    interp.MAP_START = 1
+    interp.MAP_OFFSET = 1
+    interp._join = False
+    interp._vertical_join = False
+    interp.use_encoding = False
+    interp.stack = []
 
 
 def reshape(arr, shape):
@@ -43,12 +43,12 @@ def reshape(arr, shape):
 
 
 def to_list(vector):
-    typ = interpreter.VY_type(vector)
-    if typ in (list, interpreter.Generator):
+    typ = interp.vy_type(vector)
+    if typ in (list, interp.Generator):
         return list(
             map(
                 to_list,
-                vector._dereference() if typ is interpreter.Generator else vector
+                vector._dereference() if typ is interp.Generator else vector
             )
         )
     return vector
