+++ vyxal/commands.py
@@ -46,7 +46,7 @@ command_dict = {
     "%": ("rhs, lhs = pop(stack, 2); stack.append(modulo(lhs, rhs))", 2),
     "*": ("rhs, lhs = pop(stack, 2); stack.append(multiply(lhs, rhs))", 2),
     "+": ("rhs, lhs = pop(stack, 2); stack.append(add(lhs, rhs))", 2),
-    ",": ("VY_print(pop(stack))", 1),
+    ",": ("vy_print(pop(stack))", 1),
     "-": ("rhs, lhs = pop(stack, 2); stack.append(subtract(lhs, rhs))", 2),
     "/": ("rhs, lhs = pop(stack, 2); stack.append(divide(lhs, rhs))", 2),
     ":": ("temp = pop(stack); stack.append(temp); stack.append(deref(temp))", 1),
@@ -66,30 +66,30 @@ command_dict = {
     ),
     "?": ("stack.append(get_input(0))", 0),
     "A": ("stack.append(int(all(iterable(pop(stack)))))", 1),
-    "B": ("stack.append(VY_int(pop(stack), 2))", 1),
+    "B": ("stack.append(vy_int(pop(stack), 2))", 1),
     "C": ("stack.append(chrord(pop(stack)))", 1),
     "D": (
         "temp = pop(stack); stack.append(temp); stack.append(deref(temp)); stack.append(deref(stack[-1]))",
         1,
     ),
-    "E": ("stack.append(VY_eval(pop(stack)))", 1),
-    "F": ("fn, vector = pop(stack, 2); stack.append(VY_filter(fn, vector))", 2),
-    "G": ("stack.append(VY_max(iterable(pop(stack))))", 1),
-    "H": ("stack.append(VY_int(pop(stack), 16))", 1),
-    "I": ("stack.append(VY_int(pop(stack)))", 1),
+    "E": ("stack.append(vy_eval(pop(stack)))", 1),
+    "F": ("fn, vector = pop(stack, 2); stack.append(vy_filter(fn, vector))", 2),
+    "G": ("stack.append(vy_max(iterable(pop(stack))))", 1),
+    "H": ("stack.append(vy_int(pop(stack), 16))", 1),
+    "I": ("stack.append(vy_int(pop(stack)))", 1),
     "J": ("rhs, lhs = pop(stack, 2); stack.append(join(lhs, rhs))", 2),
     "K": ("stack.append(divisors_of(pop(stack)))", 1),
     "L": ("top = pop(stack); stack.append(len(iterable(top)))", 1),
-    "M": ("fn, vector = pop(stack, 2); stack.append(VY_map(fn, vector))", 2),
+    "M": ("fn, vector = pop(stack, 2); stack.append(vy_map(fn, vector))", 2),
     "N": ("stack.append(negate(pop(stack)))", 1),
     "O": (
         "needle, haystack = pop(stack, 2); stack.append(iterable(haystack).count(needle))",
         2,
     ),
-    "P": ("rhs, lhs = pop(stack, 2); stack.append(VY_str(lhs).strip(VY_str(rhs)))", 2),
+    "P": ("rhs, lhs = pop(stack, 2); stack.append(vy_str(lhs).strip(vy_str(rhs)))", 2),
     "Q": ("exit()", 0),
-    "R": ("fn, vector = pop(stack, 2); stack += VY_reduce(fn, vector)", 2),
-    "S": ("stack.append(VY_str(pop(stack)))", 1),
+    "R": ("fn, vector = pop(stack, 2); stack += vy_reduce(fn, vector)", 2),
+    "S": ("stack.append(vy_str(pop(stack)))", 1),
     "T": ("stack.append([i for (i, x) in enumerate(pop(stack)) if bool(x)])", 1),
     "U": ("stack.append(Generator(uniquify(pop(stack))))", 1),
     "V": (
@@ -100,19 +100,19 @@ command_dict = {
     "X": ("context_level += 1", 0),
     "Y": ("rhs, lhs = pop(stack, 2); stack.append(interleave(lhs, rhs))", 2),
     "Z": (
-        "rhs, lhs = pop(stack, 2); stack.append(Generator(VY_zip(iterable(lhs), iterable(rhs))))",
+        "rhs, lhs = pop(stack, 2); stack.append(Generator(vy_zip(iterable(lhs), iterable(rhs))))",
         2,
     ),
     "a": ("stack.append(int(any(iterable(pop(stack)))))", 1),
-    "b": ("stack.append(VY_bin(pop(stack)))", 1),
+    "b": ("stack.append(vy_bin(pop(stack)))", 1),
     "c": (
-        "needle, haystack = pop(stack, 2); haystack = iterable(haystack, str)\nif type(haystack) is str: needle = VY_str(needle)\nstack.append(int(needle in iterable(haystack, str)))",
+        "needle, haystack = pop(stack, 2); haystack = iterable(haystack, str)\nif type(haystack) is str: needle = vy_str(needle)\nstack.append(int(needle in iterable(haystack, str)))",
         2,
     ),
     "d": ("stack.append(multiply(pop(stack), 2))", 1),
     "e": ("rhs, lhs = pop(stack, 2); stack.append(exponate(lhs, rhs))", 2),
     "f": ("stack.append(flatten(iterable(pop(stack))))", 1),
-    "g": ("stack.append(VY_min(iterable(pop(stack))))", 1),
+    "g": ("stack.append(vy_min(iterable(pop(stack))))", 1),
     "h": ("stack.append(iterable(pop(stack))[0])", 1),
     "i": ("rhs, lhs = pop(stack, 2)\nstack.append(index(lhs, rhs))", 2),
     "j": ("rhs, lhs = pop(stack, 2); stack.append(join_on(lhs, rhs))", 2),
@@ -124,19 +124,19 @@ command_dict = {
         2,
     ),
     "p": ("rhs, lhs = pop(stack, 2); stack.append(prepend(lhs, rhs))", 2),
-    "q": ("stack.append(uneval(VY_str(pop(stack))))", 1),
+    "q": ("stack.append(uneval(vy_str(pop(stack))))", 1),
     "r": ("rhs, lhs = pop(stack, 2); stack.append(orderless_range(lhs, rhs))", 2),
-    "s": ("stack.append(VY_sorted(pop(stack)))", 1),
+    "s": ("stack.append(vy_sorted(pop(stack)))", 1),
     "t": ("stack.append(iterable(pop(stack))[-1])", 1),
     "u": ("stack.append(-1)", 0),
     "w": ("stack.append([pop(stack)])", 1),
     "x": ("stack += this_function(stack)", 0),
     "y": ("stack += uninterleave(pop(stack))", 1),
-    "z": ("fn, vector = pop(stack, 2); stack += VY_zipmap(fn, vector)", 2),
+    "z": ("fn, vector = pop(stack, 2); stack += vy_zipmap(fn, vector)", 2),
     "↑": ("stack.append(max(pop(stack), key=lambda x: x[-1]))", 1),
     "↓": ("stack.append(min(pop(stack), key=lambda x: x[-1]))", 1),
-    "∴": ("rhs, lhs = pop(stack, 2); stack.append(VY_max(lhs, rhs))", 2),
-    "∵": ("rhs, lhs = pop(stack, 2); stack.append(VY_min(lhs, rhs))", 2),
+    "∴": ("rhs, lhs = pop(stack, 2); stack.append(vy_max(lhs, rhs))", 2),
+    "∵": ("rhs, lhs = pop(stack, 2); stack.append(vy_min(lhs, rhs))", 2),
     "β": (
         "alphabet, number = pop(stack, 2); stack.append(utilities.to_ten(number, alphabet))",
         2,
@@ -150,17 +150,17 @@ command_dict = {
     "∷": ("stack.append(modulo(pop(stack), 2))", 1),
     "¤": ("stack.append('')", 0),
     "ð": ("stack.append(' ')", 0),
-    "ȧ": ("stack.append(VY_abs(pop(stack)))", 1),
+    "ȧ": ("stack.append(vy_abs(pop(stack)))", 1),
     "ḃ": ("stack.append(int(not compare(pop(stack), 0, Comparitors.EQUALS)))", 1),
     "ċ": ("stack.append(compare(pop(stack), 1, Comparitors.NOT_EQUALS))", 1),
     "ḋ": (
-        "rhs, lhs = pop(stack, 2); stack.append(VY_divmod(lhs, rhs))",
+        "rhs, lhs = pop(stack, 2); stack.append(vy_divmod(lhs, rhs))",
         2,
     ),  # Dereference because generators could accidentally get exhausted.
     "ė": ("stack.append(Generator(enumerate(iterable(pop(stack)))))", 1),
     "ḟ": ("rhs, lhs = pop(stack, 2); stack.append(find(lhs, rhs))", 2),
     "ġ": (
-        "rhs = pop(stack)\nif VY_type(rhs) in [list, Generator]: stack.append(gcd(rhs))\nelse: stack.append(gcd(pop(stack), rhs))",
+        "rhs = pop(stack)\nif vy_type(rhs) in [list, Generator]: stack.append(gcd(rhs))\nelse: stack.append(gcd(pop(stack), rhs))",
         2,
     ),
     "ḣ": ("top = iterable(pop(stack)); stack.append(top[0]); stack.append(top[1:])", 1),
@@ -176,15 +176,15 @@ command_dict = {
     "ṅ": ("stack.append(first_n(pop(stack)))", 1),
     "ȯ": ("n, fn = pop(stack, 2); stack.append(first_n(fn, n))", 2),
     "ṗ": ("stack.append(powerset(iterable(pop(stack))))", 1),
-    "ṙ": ("stack.append(VY_round(pop(stack)))", 1),
-    "ṡ": ("fn , vector = pop(stack, 2); stack.append(VY_sorted(vector, fn))", 2),
+    "ṙ": ("stack.append(vy_round(pop(stack)))", 1),
+    "ṡ": ("fn , vector = pop(stack, 2); stack.append(vy_sorted(vector, fn))", 2),
     "ṫ": (
         "vector = iterable(pop(stack)); stack.append(vector[:-1]); stack.append(vector[-1])",
         1,
     ),
     "ẇ": ("rhs, lhs = pop(stack, 2); stack.append(wrap(lhs, rhs))", 2),
     "ẋ": (
-        "rhs, lhs = pop(stack, 2); main = None;\nif VY_type(lhs) is Function: main = pop(stack)\nstack.append(repeat(lhs, rhs, main))",
+        "rhs, lhs = pop(stack, 2); main = None;\nif vy_type(lhs) is Function: main = pop(stack)\nstack.append(repeat(lhs, rhs, main))",
         2,
     ),
     "ẏ": ("obj = iterable(pop(stack)); stack.append(Generator(range(0, len(obj))))", 1),
@@ -232,10 +232,10 @@ command_dict = {
         "rhs, lhs = pop(stack, 2); ret = is_divisble(lhs, rhs)\nif type(ret) is tuple: stack += list(ret)\nelse: stack.append(ret)",
         2,
     ),
-    "Ė": ("stack += VY_exec(pop(stack))", 1),
+    "Ė": ("stack += vy_exec(pop(stack))", 1),
     "Ḟ": (
         """top = pop(stack)
-if VY_type(top) is Number:
+if vy_type(top) is Number:
     limit = int(top); vector = pop(stack)
 else:
     limit = -1; vector = top
@@ -259,7 +259,7 @@ stack.append(Generator(fn, limit=limit, initial=iterable(vector)))
         3,
     ),
     "Ṅ": (
-        "top = pop(stack);\nif VY_type(top) == Number:stack.append(Generator(partition(top)))\nelse: stack.append(' '.join([VY_str(x) for x in top]))",
+        "top = pop(stack);\nif vy_type(top) == Number:stack.append(Generator(partition(top)))\nelse: stack.append(' '.join([vy_str(x) for x in top]))",
         1,
     ),  # ---------------------------
     "Ȯ": (
@@ -291,10 +291,10 @@ stack.append(Generator(fn, limit=limit, initial=iterable(vector)))
     "⌊": ("stack.append(floor(pop(stack)))", 1),
     "¯": ("stack.append(deltas(pop(stack)))", 1),
     "±": ("stack.append(sign_of(pop(stack)))", 1),
-    "₴": ("VY_print(pop(stack), end='')", 1),
-    "…": ("top = pop(stack); stack.append(top); VY_print(top)", 0),
+    "₴": ("vy_print(pop(stack), end='')", 1),
+    "…": ("top = pop(stack); stack.append(top); vy_print(top)", 0),
     "□": (
-        "if inputs: stack.append(inputs)\nelse:\n    s, x = [], input()\n    while x:\n        s.append(Vy_eval(x)); x = input()",
+        "if inputs: stack.append(inputs)\nelse:\n    s, x = [], input()\n    while x:\n        s.append(vy_eval(x)); x = input()",
         0,
     ),
     "↳": ("rhs, lhs = pop(stack, 2); stack.append(rshift(lhs, rhs))", 2),
@@ -386,7 +386,7 @@ stack.append(Generator(fn, limit=limit, initial=iterable(vector)))
     "∆D": ("arg = pop(stack); stack.append(vectorise(math.degrees, arg))", 1),
     "∆R": ("arg = pop(stack); stack.append(vectorise(math.radians, arg))", 1),
     "∆≤": (
-        "arg = pop(stack); stack.append(compare(VY_abs(arg), 1, Comparitors.LESS_THAN_EQUALS))",
+        "arg = pop(stack); stack.append(compare(vy_abs(arg), 1, Comparitors.LESS_THAN_EQUALS))",
         1,
     ),
     "∆Ṗ": ("stack.append(next_prime(pop(stack)))", 1),
@@ -426,9 +426,9 @@ stack.append(Generator(fn, limit=limit, initial=iterable(vector)))
     "øe": ("stack.append(run_length_encode(iterable(pop(stack), str)))", 1),
     "ød": ("stack.append(run_length_decode(pop(stack)))", 1),
     "øD": ("stack.append(dictionary_compress(pop(stack)))", 1),
-    "øW": ("stack.append(split_on_words(VY_str(pop(stack))))", 1),
+    "øW": ("stack.append(split_on_words(vy_str(pop(stack))))", 1),
     "øṙ": (
-        "replacent, pattern, source = pop(stack, 3); stack.append(regex_replace(VY_str(source), VY_str(pattern), replacent))",
+        "replacent, pattern, source = pop(stack, 3); stack.append(regex_replace(vy_str(source), vy_str(pattern), replacent))",
         3,
     ),
     "øp": (
@@ -444,11 +444,11 @@ stack.append(Generator(fn, limit=limit, initial=iterable(vector)))
     "ø¦": ("rhs, lhs = pop(stack, 2); stack.append(vertical_mirror(lhs, rhs))", 2),
     "Þ…": ("value, vector = pop(stack, 2); stack.append(distribute(vector, value))", 2),
     "Þ↓": (
-        "fn, vector = pop(stack, 2); stack.append(min(VY_zipmap(fn, vector), key=lambda x: x[-1])[0])",
+        "fn, vector = pop(stack, 2); stack.append(min(vy_zipmap(fn, vector), key=lambda x: x[-1])[0])",
         2,
     ),
     "Þ↑": (
-        "fn, vector = pop(stack, 2); stack.append(max(VY_zipmap(fn, vector), key=lambda x: x[-1])[0])",
+        "fn, vector = pop(stack, 2); stack.append(max(vy_zipmap(fn, vector), key=lambda x: x[-1])[0])",
         2,
     ),
     "Þ×": ("vector = pop(stack); stack.append(all_combinations(vector));", 1),
@@ -487,8 +487,8 @@ stack.append(Generator(fn, limit=limit, initial=iterable(vector)))
         "function, indexes, original = pop(stack, 3); stack.append(map_at(function, iterable(original), iterable(indexes)))",
         3,
     ),
-    "¨,": ("VY_print(pop(stack), end=' ')", 1),
-    "¨…": ("top = pop(stack); stack.append(top); VY_print(top, end=' ')", 1),
+    "¨,": ("vy_print(pop(stack), end=' ')", 1),
+    "¨…": ("top = pop(stack); stack.append(top); vy_print(top, end=' ')", 1),
     "¨t": ("vectorise(time.sleep, pop(stack))", 1),
     "kA": ("stack.append(string.ascii_uppercase)", 0),
     "ke": ("stack.append(math.e)", 0),
