+++ vyxal/utilities.py
@@ -1,4 +1,271 @@
 from vyxal import words
+from vyxal.globals import *
+
+# Generic type constants
+Number = "NUMBER"
+Iterable = "ITERABLE"
+Function = type(lambda: None)
+
+Python_Generator = type(
+    i for i in (0,)
+)  # https://chat.stackexchange.com/transcript/message/57555979#57555979
+
+NEWLINE = "\n"
+ONE_TWO_EIGHT_KB = 1024000
+base53alphabet = "¡etaoinshrdlcumwfgypbvkjxqz ETAOINSHRDLCUMWFGYPBVKJXQZ"
+base27alphabet = " etaoinshrdlcumwfgypbvkjxqz"
+
+# Helper classes
+class Comparitors:
+    EQUALS = 0
+    LESS_THAN = 1
+    GREATER_THAN = 2
+    NOT_EQUALS = 3
+    LESS_THAN_EQUALS = 4
+    GREATER_THAN_EQUALS = 5
+
+
+class Generator:
+    def __init__(
+        self,
+        raw_generator,
+        limit=-1,
+        initial=[],
+        condition=None,
+        is_numeric_sequence=False,
+    ):
+        self.next_index = 0
+        self.end_reached = False
+        self.is_numeric_sequence = is_numeric_sequence
+        self.do_print = True
+        if "__name__" in dir(raw_generator) and type(raw_generator) != Python_Generator:
+            if raw_generator.__name__.startswith(
+                "FN_"
+            ) or raw_generator.__name__.startswith("_lambda"):
+                # User defined function
+                def gen():
+                    generated = initial
+                    factor = len(initial)
+                    for item in initial:
+                        yield item
+                    while True:
+                        if len(generated) >= (limit + factor) and limit > 0:
+                            break
+                        else:
+                            ret = raw_generator(generated[::-1], arity=len(generated))
+                            generated.append(ret[-1])
+                            yield ret[-1]
+
+                self.gen = gen()
+            else:
+
+                def gen():
+                    index = 0
+                    while True:
+                        yield raw_generator(index)
+                        index += 1
+
+                self.gen = gen()
+        else:
+
+            def niceify(item):
+                t_item = vy_type(item)
+                if t_item not in [Generator, list, Number, str]:
+                    return list(item)
+                return item
+
+            self.gen = map(niceify, raw_generator)
+        self.generated = []
+
+    def __contains__(self, item):
+        if self.is_numeric_sequence:
+            if item in self.generated:
+                return True
+            temp = next(self)
+            while temp <= item:
+                temp = next(self)
+            return item in self.generated
+        else:
+            for temp in self:
+                if item == temp:
+                    return True
+            return False
+
+    def __getitem__(self, position):
+        if type(position) is slice:
+            ret = []
+            stop = position.stop or self.__len__()
+            start = position.start or 0
+
+            if start < 0 or stop < 0:
+                self.generated += list(self.gen)
+                return self.generated[position]
+            if stop < 0:
+                stop = self.__len__() - position.stop - 2
+
+            if position.step and position.step < 0:
+                start, stop = stop, start
+                stop -= 1
+                start -= 1
+            # print(start, stop, position.step or 1)
+            for i in range(start, stop, position.step or 1):
+                ret.append(self.__getitem__(i))
+                # print(self.__getitem__(i))
+            return ret
+        if position < 0:
+            self.generated += list(self.gen)
+            return self.generated[position]
+        if position < len(self.generated):
+            return self.generated[position]
+        while len(self.generated) < position + 1:
+            try:
+                self.__next__()
+            except:
+                self.end_reached = True
+                position = position % len(self.generated)
+
+        return self.generated[position]
+
+    def __setitem__(self, position, value):
+        if position >= len(self.generated):
+            temp = self.__getitem__(position)
+        self.generated[position] = value
+
+    def __len__(self):
+        return len(self._dereference())
+
+    def __next__(self):
+        f = next(self.gen)
+        self.generated.append(f)
+        return f
+
+    def __iter__(self):
+        return self
+
+    def _filter(self, function):
+        index = 0
+        l = self.__len__()
+        while True:
+            if index == l:
+                break
+            obj = self.__getitem__(index)
+            ret = _safe_apply(function, obj)
+            if ret:
+                yield obj
+            index += 1
+
+    def _reduce(self, function):
+        def ensure_singleton(function, left, right):
+            ret = _safe_apply(function, left, right)
+            if type(ret) in [Generator, list]:
+                return ret[-1]
+            return ret
+
+        return functools.reduce(
+            lambda x, y: ensure_singleton(function, x, y), self._dereference()
+        )
+
+    def _dereference(self):
+        """
+        Only call this when it is absolutely neccesary to convert to a list.
+        """
+        d = self.generated + list(self.gen)
+        self.gen = iter(d[::])
+        self.generated = []
+        return d
+
+    def _print(self, end="\n"):
+        main = self.generated
+        try:
+            f = next(self)
+            # If we're still going, there's stuff in main that needs printing before printing the generator
+            vy_print("⟨", end="")
+            for i in range(len(main)):
+                vy_print(main[i], end="|" * (i >= len(main)))
+            while True:
+                try:
+                    f = next(self)
+                    vy_print("|", end="")
+                    vy_print(f, end="")
+                except:
+                    break
+            vy_print("⟩", end=end)
+
+        except:
+            vy_print(main, end=end)
+
+    def zip_with(self, other):
+        return Generator(zip(self.gen, iter(other)))
+
+    def safe(self):
+        import copy
+
+        return copy.deepcopy(self)
+
+    def __str__(self):
+        return "⟨" + "|".join(str(item for item in self.generated)) + "...⟩"
+
+    def limit_to_items(self, n):
+        out = "⟨"
+        item_count = 0
+        while not self.end_reached and item_count <= n:
+            item = self.__getitem__(item_count)
+            if self.end_reached:
+                break
+            out += (
+                str(item) if vy_type(item) is not Generator else item.limit_to_items(n)
+            )
+            item_count += 1
+            out += "|"
+
+        if item_count > n:
+            out += "..."
+
+        return out + "⟩"
+
+
+class ShiftDirections:
+    LEFT = 1
+    RIGHT = 2
+
+
+# Helper functions
+def _safe_apply(function, *args):
+    """
+    Applies function to args that adapts to the input style of the passed function.
+
+    If the function is a _lambda (it's been defined within λ...;), it passes a list of arguments and length of argument list
+    Otherwise, if the function is a user-defined function (starts with FN_), it simply passes the argument list
+    Otherwise, unpack args and call as usual
+    """
+
+    if function.__name__.startswith("_lambda"):
+        ret = function(list(args), len(args), function)
+        if len(ret):
+            return ret[-1]
+        else:
+            return []
+    elif function.__name__.startswith("FN_"):
+        ret = function(list(args))[-1]
+        if len(ret):
+            return ret[-1]
+        else:
+            return []
+    return function(*args)
+
+
+def _mangle(value):
+    byte_list = bytes(value, encoding="utf-8")
+    return base64.b32encode(byte_list).decode().replace("=", "_")
+
+
+def _two_argument(function, left, right):
+    """
+    Used for vectorising user-defined lambas/dyads over generators
+    """
+    if function.__name__.startswith("_lambda"):
+        return Generator(map(lambda x: function(x, arity=2), vy_zip(left, right)))
+    return Generator(map(lambda x: function(*x), vy_zip(left, right)))
 
 
 def to_ten(number, custom_base):
@@ -102,8 +369,111 @@ def uncompress(s):
     return final.replace("\n", "\\n").replace("\r", "")
 
 
-base53alphabet = "¡etaoinshrdlcumwfgypbvkjxqz ETAOINSHRDLCUMWFGYPBVKJXQZ"
-base27alphabet = " etaoinshrdlcumwfgypbvkjxqz"
+def vy_print(item, end="\n", raw=False):
+    global output, printed, stack
+    printed = True
+    t_item = type(item)
+    if t_item is Generator:
+        item._print(end)
+
+    elif t_item is list:
+        vy_print("⟨", "", False)
+        if item:
+            for value in item[:-1]:
+                vy_print(value, "|", True)
+            vy_print(item[-1], "", True)
+        vy_print("⟩", end, False)
+    elif t_item is Function:
+        pop(stack)
+        s = function_call(item, stack)
+        vy_print(s[0], end=end, raw=raw)
+    else:
+        if t_item is int and keg_mode:
+            item = chr(item)
+        if raw:
+            if online_version:
+                output[1] += vy_repr(item) + end
+            else:
+                print(vy_repr(item), end=end)
+        else:
+            if online_version:
+                output[1] += vy_str(item) + end
+            else:
+                print(vy_str(item), end=end)
+    if online_version and len(output) > ONE_TWO_EIGHT_KB:
+        exit(code=1)
+
+
+def vy_str(item):
+    global stack
+    t_item = vy_type(item)
+    return {
+        Number: str,
+        str: lambda x: x,
+        list: lambda x: "⟨" + "|".join([vy_repr(y) for y in x]) + "⟩",
+        Generator: lambda x: vy_str(x._dereference()),
+        Function: lambda x: vy_str(function_call(item, stack)[0]),
+    }[t_item](item)
+
+
+def vy_repr(item):
+    t_item = vy_type(item)
+    return {
+        Number: str,
+        list: lambda x: "⟨" + "|".join([str(vy_repr(y)) for y in x]) + "⟩",
+        Generator: lambda x: vy_repr(x._dereference()),
+        str: lambda x: "`" + x + "`",
+        Function: lambda x: "@FUNCTION:" + x.__name__,
+    }[t_item](item)
+
+
+def vy_type(item):
+    ty = type(item)
+    if ty in [int, float, complex]:
+        return Number
+    return ty
+
+
+def vy_zip(lhs, rhs):
+    ind = 0
+    if type(lhs) in [list, str]:
+        lhs = iter(lhs)
+    if type(rhs) in [list, str]:
+        rhs = iter(rhs)
+    while True:
+        exhausted = 0
+        try:
+            l = next(lhs)
+        except:
+            l = 0
+            exhausted += 1
+
+        try:
+            r = next(rhs)
+        except:
+            r = 0
+            exhausted += 1
+        if exhausted == 2:
+            break
+        yield [l, r]
+        ind += 1
+
+
+def vy_zipmap(lhs, rhs):
+    if Function not in (vy_type(lhs), vy_type(rhs)):
+        return [lhs, vy_zip(rhs, rhs)]
+
+    function = vector = None
+    if type(rhs) is Function:
+        function, vector = rhs, iterable(lhs, range)
+    else:
+        function, vector = lhs, iterable(rhs, range)
+
+    def f():
+        for item in vector:
+            yield [item, _safe_apply(function, item)]
+
+    return Generator(f())
 
 
 if __name__ == "__main__":
